+++ packages/preview-middleware-client/src/adp/quick-actions/common/create-page-action.ts
@@ -27,7 +27,7 @@ export class AddPageActionQuickAction extends SimpleQuickActionDefinitionBase im
         if (appType === 'fe-v4' && isLowerThanMinimalUi5Version(version, { major: 1, minor: 130 })) {
             return;
         }
-        super.initialize();
+        await super.initialize();
     }
 
     async execute(): Promise<FlexCommand[]> {