+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/fe-fpm-cli
 
+## 0.0.65
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.33.5
+
 ## 0.0.64
 
 ### Patch Changes