+++ packages/cf-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/cf-deploy-config-sub-generator
 
+## 0.1.9
+
+### Patch Changes
+
+-   Updated dependencies [c50e09f]
+    -   @sap-ux/project-access@1.29.8
+    -   @sap-ux/cf-deploy-config-writer@0.1.11
+    -   @sap-ux/fiori-generator-shared@0.8.1
+    -   @sap-ux/deploy-config-generator-shared@0.0.25
+    -   @sap-ux/inquirer-common@0.6.18
+    -   @sap-ux/cf-deploy-config-inquirer@0.2.6
+
 ## 0.1.8
 
 ### Patch Changes