+++ packages/ui5-application-inquirer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/ui5-application-inquirer
 
+## 0.9.18
+
+### Patch Changes
+
+-   Updated dependencies [c50e09f]
+    -   @sap-ux/project-access@1.29.8
+    -   @sap-ux/inquirer-common@0.6.18
+
 ## 0.9.17
 
 ### Patch Changes